Move services' tests classes to correct package
