+++ Kitodo/src/test/java/org/kitodo/services/data/TemplatePropertyServiceIT.java
@@ -9,7 +9,7 @@
  * GPL3-License.txt file that was distributed with this source code.
  */
 
-package org.kitodo.services;
+package org.kitodo.services.data;
 
 import java.io.IOException;
 import java.util.List;
